requirements.txt
streamlit
pandas
numpy
scikit-learn
sentence-transformers
faiss-cpu
pdfplumber
matplotlib
seaborn
openai
python-dotenv